+++ ChangeLog.md
@@ -13,6 +13,7 @@
 * Appli mobile : Changements de certains libellés pour LRDS.
 * Tags / onglet Collecte : correctif de l'écran. Certains libellés n'étaient pas affichés, ou étaient incorrect.
 * Nettoyage du code (diverses modifications).
+* Ajout de StateVirtual.
 
 ## 0.12